4526 - move countryStatuses under area
